Move system clock configuration routines out of main.c to reduce clutter
