+++ examples/stm32/xively_demo_stm32f4xx_nucleo_wifi/Inc/demo_bsp.h
@@ -1,0 +1,23 @@
+#ifndef __DEMO_BSP__
+#define __DEMO_BSP__
+
+#ifdef USE_STM32L0XX_NUCLEO
+#include "stm32l0xx_hal.h"
+#include "stm32l0xx_nucleo.h"
+#endif
+#ifdef USE_STM32F1xx_NUCLEO
+#include "stm32f1xx_hal.h"
+#include "stm32f1xx_nucleo.h"
+#endif
+#ifdef USE_STM32F4XX_NUCLEO
+#include "stm32f4xx_hal.h"
+#include "stm32f4xx_nucleo.h"
+#endif  
+#ifdef USE_STM32L4XX_NUCLEO
+#include "stm32l4xx_hal.h"
+#include "stm32l4xx_nucleo.h"
+#endif 
+
+void SystemClock_Config( void );
+
+#endif /* __DEMO_BSP__ */
