+++ examples/stm32/xively_demo_stm32f4xx_nucleo_wifi/Src/main.c
@@ -40,6 +40,7 @@
 #include "wifi_interface.h"
 #include "stdio.h"
 #include "string.h"
+#include "demo_bsp.h"
 
 /* Xively's headers */
 #include "xively.h"
@@ -492,6 +493,19 @@ static inline int8_t system_init( void )
         printf( "Error in Config" );
         return -1;
     }
+    return 0;
+}
+
+void flash_read( void )
+{
+    HAL_FLASH_Lock();
+    HAL_FLASH_Unlock();
+}
+
+void flash_write( void )
+{
+    HAL_FLASH_Lock();
+    HAL_FLASH_Unlock();
 }
 
 /**
@@ -623,194 +637,6 @@ void HAL_GPIO_EXTI_Callback( uint16_t GPIO_Pin )
         }
     }
 }
-
-/**
- * @brief  System Clock Configuration
- *         The system Clock is configured as follow :
- *            System Clock source            = PLL (HSI)
- *            SYSCLK(Hz)                     = 64000000
- *            HCLK(Hz)                       = 64000000
- *            AHB Prescaler                  = 1
- *            APB1 Prescaler                 = 2
- *            APB2 Prescaler                 = 1
- *            PLLMUL                         = 16
- *            Flash Latency(WS)              = 2
- * @param  None
- * @retval None
- */
-
-#ifdef USE_STM32F1xx_NUCLEO
-
-void SystemClock_Config( void )
-{
-    RCC_ClkInitTypeDef clkinitstruct = {0};
-    RCC_OscInitTypeDef oscinitstruct = {0};
-
-    /* Configure PLL ------------------------------------------------------*/
-    /* PLL configuration: PLLCLK = (HSI / 2) * PLLMUL = (8 / 2) * 16 = 64 MHz */
-    /* PREDIV1 configuration: PREDIV1CLK = PLLCLK / HSEPredivValue = 64 / 1 = 64 MHz */
-    /* Enable HSI and activate PLL with HSi_DIV2 as source */
-    oscinitstruct.OscillatorType      = RCC_OSCILLATORTYPE_HSE;
-    oscinitstruct.HSEState            = RCC_HSE_ON;
-    oscinitstruct.LSEState            = RCC_LSE_OFF;
-    oscinitstruct.HSIState            = RCC_HSI_OFF;
-    oscinitstruct.HSICalibrationValue = RCC_HSICALIBRATION_DEFAULT;
-    oscinitstruct.HSEPredivValue      = RCC_HSE_PREDIV_DIV1;
-    oscinitstruct.PLL.PLLState        = RCC_PLL_ON;
-    oscinitstruct.PLL.PLLSource       = RCC_PLLSOURCE_HSE;
-    oscinitstruct.PLL.PLLMUL          = RCC_PLL_MUL9;
-    if ( HAL_RCC_OscConfig( &oscinitstruct ) != HAL_OK )
-    {
-        /* Initialization Error */
-        while ( 1 )
-            ;
-    }
-
-    /* Select PLL as system clock source and configure the HCLK, PCLK1 and PCLK2
-       clocks dividers */
-    clkinitstruct.ClockType = ( RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_HCLK |
-                                RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2 );
-    clkinitstruct.SYSCLKSource   = RCC_SYSCLKSOURCE_PLLCLK;
-    clkinitstruct.AHBCLKDivider  = RCC_SYSCLK_DIV1;
-    clkinitstruct.APB2CLKDivider = RCC_HCLK_DIV1;
-    clkinitstruct.APB1CLKDivider = RCC_HCLK_DIV2;
-    if ( HAL_RCC_ClockConfig( &clkinitstruct, FLASH_LATENCY_2 ) != HAL_OK )
-    {
-        /* Initialization Error */
-        while ( 1 )
-            ;
-    }
-}
-#endif
-
-#ifdef USE_STM32F4XX_NUCLEO
-
-void SystemClock_Config( void )
-{
-    RCC_ClkInitTypeDef RCC_ClkInitStruct;
-    RCC_OscInitTypeDef RCC_OscInitStruct;
-
-    /* Enable Power Control clock */
-    __PWR_CLK_ENABLE();
-
-    /* The voltage scaling allows optimizing the power consumption when the device is
-       clocked below the maximum system frequency, to update the voltage scaling value
-       regarding system frequency refer to product datasheet.  */
-    __HAL_PWR_VOLTAGESCALING_CONFIG( PWR_REGULATOR_VOLTAGE_SCALE2 );
-
-    /* Enable HSI Oscillator and activate PLL with HSI as source */
-    RCC_OscInitStruct.OscillatorType      = RCC_OSCILLATORTYPE_HSI;
-    RCC_OscInitStruct.HSIState            = RCC_HSI_ON;
-    RCC_OscInitStruct.HSICalibrationValue = 0x10;
-    RCC_OscInitStruct.PLL.PLLState        = RCC_PLL_ON;
-    RCC_OscInitStruct.PLL.PLLSource       = RCC_PLLSOURCE_HSI;
-    RCC_OscInitStruct.PLL.PLLM            = 16;
-    RCC_OscInitStruct.PLL.PLLN            = 336;
-    RCC_OscInitStruct.PLL.PLLP            = RCC_PLLP_DIV4;
-    RCC_OscInitStruct.PLL.PLLQ            = 7;
-    HAL_RCC_OscConfig( &RCC_OscInitStruct );
-
-    /* Select PLL as system clock source and configure the HCLK, PCLK1 and PCLK2
-       clocks dividers */
-    RCC_ClkInitStruct.ClockType = ( RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_HCLK |
-                                    RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2 );
-    RCC_ClkInitStruct.SYSCLKSource   = RCC_SYSCLKSOURCE_PLLCLK;
-    RCC_ClkInitStruct.AHBCLKDivider  = RCC_SYSCLK_DIV1;
-    RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV2;
-    RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-    HAL_RCC_ClockConfig( &RCC_ClkInitStruct, FLASH_LATENCY_2 );
-}
-#endif
-
-#ifdef USE_STM32L0XX_NUCLEO
-
-/**
- * @brief  System Clock Configuration
- * @param  None
- * @retval None
- */
-void SystemClock_Config( void )
-{
-    RCC_ClkInitTypeDef RCC_ClkInitStruct;
-    RCC_OscInitTypeDef RCC_OscInitStruct;
-
-    __PWR_CLK_ENABLE();
-
-    __HAL_PWR_VOLTAGESCALING_CONFIG( PWR_REGULATOR_VOLTAGE_SCALE1 );
-
-    RCC_OscInitStruct.OscillatorType      = RCC_OSCILLATORTYPE_HSI;
-    RCC_OscInitStruct.HSIState            = RCC_HSI_ON;
-    RCC_OscInitStruct.HSICalibrationValue = 0x10;
-    RCC_OscInitStruct.PLL.PLLState        = RCC_PLL_ON;
-    RCC_OscInitStruct.PLL.PLLSource       = RCC_PLLSOURCE_HSI;
-    RCC_OscInitStruct.PLL.PLLMUL          = RCC_PLLMUL_4;
-    RCC_OscInitStruct.PLL.PLLDIV          = RCC_PLLDIV_2;
-    HAL_RCC_OscConfig( &RCC_OscInitStruct );
-
-    RCC_ClkInitStruct.ClockType =
-        ( RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_HCLK | RCC_CLOCKTYPE_PCLK1 |
-          RCC_CLOCKTYPE_PCLK2 );                             // RCC_CLOCKTYPE_SYSCLK;
-    RCC_ClkInitStruct.SYSCLKSource   = RCC_SYSCLKSOURCE_HSI; // RCC_SYSCLKSOURCE_PLLCLK;
-    RCC_ClkInitStruct.AHBCLKDivider  = RCC_SYSCLK_DIV1;
-    RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV2;
-    RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-    HAL_RCC_ClockConfig( &RCC_ClkInitStruct, FLASH_LATENCY_1 );
-
-    __SYSCFG_CLK_ENABLE();
-}
-#endif
-
-#ifdef USE_STM32L4XX_NUCLEO
-/**
- * @brief  System Clock Configuration
- *         The system Clock is configured as follow :
- *            System Clock source            = PLL (MSI)
- *            SYSCLK(Hz)                     = 80000000
- *            HCLK(Hz)                       = 80000000
- *            AHB Prescaler                  = 1
- *            APB1 Prescaler                 = 1
- *            APB2 Prescaler                 = 1
- *            MSI Frequency(Hz)              = 4000000
- *            PLL_M                          = 1
- *            PLL_N                          = 40
- *            PLL_R                          = 2
- *            PLL_P                          = 7
- *            PLL_Q                          = 4
- *            Flash Latency(WS)              = 4
- * @param  None
- * @retval None
- */
-void SystemClock_Config( void )
-{
-    RCC_ClkInitTypeDef RCC_ClkInitStruct = {0};
-    RCC_OscInitTypeDef RCC_OscInitStruct = {0};
-
-    /* MSI is enabled after System reset, activate PLL with MSI as source */
-    RCC_OscInitStruct.OscillatorType      = RCC_OSCILLATORTYPE_MSI;
-    RCC_OscInitStruct.MSIState            = RCC_MSI_ON;
-    RCC_OscInitStruct.MSIClockRange       = RCC_MSIRANGE_6;
-    RCC_OscInitStruct.MSICalibrationValue = RCC_MSICALIBRATION_DEFAULT;
-    RCC_OscInitStruct.PLL.PLLState        = RCC_PLL_ON;
-    RCC_OscInitStruct.PLL.PLLSource       = RCC_PLLSOURCE_MSI;
-    RCC_OscInitStruct.PLL.PLLM            = 1;
-    RCC_OscInitStruct.PLL.PLLN            = 40;
-    RCC_OscInitStruct.PLL.PLLR            = 2;
-    RCC_OscInitStruct.PLL.PLLP            = 7;
-    RCC_OscInitStruct.PLL.PLLQ            = 4;
-    HAL_RCC_OscConfig( &RCC_OscInitStruct );
-
-    /* Select PLL as system clock source and configure the HCLK, PCLK1 and PCLK2
-       clocks dividers */
-    RCC_ClkInitStruct.ClockType = ( RCC_CLOCKTYPE_SYSCLK | RCC_CLOCKTYPE_HCLK |
-                                    RCC_CLOCKTYPE_PCLK1 | RCC_CLOCKTYPE_PCLK2 );
-    RCC_ClkInitStruct.SYSCLKSource   = RCC_SYSCLKSOURCE_PLLCLK;
-    RCC_ClkInitStruct.AHBCLKDivider  = RCC_SYSCLK_DIV1;
-    RCC_ClkInitStruct.APB1CLKDivider = RCC_HCLK_DIV1;
-    RCC_ClkInitStruct.APB2CLKDivider = RCC_HCLK_DIV1;
-    HAL_RCC_ClockConfig( &RCC_ClkInitStruct, FLASH_LATENCY_4 );
-}
-#endif
-
 #ifdef USE_FULL_ASSERT
 
 /**
